Implemented jquery and controller method to allow adding books to WantToRead table

+++ src/main/java/com/example/ISTE240_Project/Services/BookService.java
@@ -23,7 +23,7 @@ public class BookService {
         return repo.findAll();
     }
 
-    public Book getBookById(Long id) {
+    public Book getBookById(int id) {
         return repo.findById(id).orElse(null);
     }
